requirements.txt
dj-database-url==0.4.1
Django==1.11.28
gunicorn==19.6.0
psycopg2==2.7.3.2
whitenoise==2.0.6
geopy>=1.11.0
django-ipware>=1.1.6
git+https://github.com/jschneier/django-storages.git#egg=django-storages
Pillow>=4.0.0
dropbox>=8.2.0
requests>=2.18.1]
six>=1.12.0